Fix first time branch errors
I found this error a little misleading. It would always error out with:

```
java.lang.NullPointerException: Cannot get property 'result' on null object
```

This was because on a fresh branch, I was trying to access
currentBuild.previousBuild.result, and currentBuild.previousBuild was
null since it was a brand new object.

This catches this exception to ensure that if it's null, we don't care
about the previous build, because, well, there is no previous build.