Add channel parameter and regressions
This changes the function slightly so rather than passing in the
slightly opaque currentBuild.result, it uses that directly. Instead we
have to pass in (since we _do_ have to pass in a parameter) the name of
the channel, which makes sense since different teams may want to notify
different channels, and this gives them the flexibility to do so.

It also adds in the idea of a "REGRESSION", which is a build that was
previously good has now failed.

I've ensured when "alertEveryone" is set to true, it won't fail due to
the currentBuild.previousBuild bug.

I've updated the emoji slightly for FUN. I LOVE FUN.